Make OpenShiftMetaEndpointIT use RHBK on OCP
* Just like all the other OCP tests using KC. For this, we need to
  extract the test logic into abstract parent also
